+++ tests/supernotify/test_notify_snooze.py
@@ -1,3 +1,5 @@
+from datetime import timedelta
+
 from homeassistant.const import CONF_EMAIL
 from homeassistant.core import Context, Event, HomeAssistant
 
@@ -37,47 +39,65 @@ DELIVERY: dict[str, dict] = {
 def test_snooze_delivery(mock_hass: HomeAssistant) -> None:
     uut = SupernotifyAction(mock_hass)
 
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_DELIVERY_foo"}))
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_DELIVERY_foo"}))
     assert list(uut.context.snoozer.snoozes.values()) == [
-        Snooze(QualifiedTargetType.DELIVERY, RecipientType.EVERYONE, "foo", snooze_for=3600)
+        Snooze(QualifiedTargetType.DELIVERY, RecipientType.EVERYONE,
+               "foo", snooze_for=timedelta(hours=1))
     ]
     assert all(s["target"] == "foo" for s in uut.enquire_snoozes())
     assert all(
-        s.snooze_until is not None and s.snooze_until - s.snoozed_at == 3600 for s in uut.context.snoozer.snoozes.values()
+        s.snooze_until is not None and s.snooze_until -
+        s.snoozed_at == timedelta(hours=1)
+        for s in uut.context.snoozer.snoozes.values()
     )
 
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SILENCE_EVERYONE_DELIVERY_foo"}))
-    assert list(uut.context.snoozer.snoozes.values()) == [Snooze(QualifiedTargetType.DELIVERY, RecipientType.EVERYONE, "foo")]
-    assert all(s.snooze_until is None for s in uut.context.snoozer.snoozes.values())
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_SILENCE_EVERYONE_DELIVERY_foo"}))
+    assert list(uut.context.snoozer.snoozes.values()) == [Snooze(
+        QualifiedTargetType.DELIVERY, RecipientType.EVERYONE, "foo")]
+    assert all(
+        s.snooze_until is None for s in uut.context.snoozer.snoozes.values())
 
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_DELIVERY_foo_33"}))
-    assert list(uut.context.snoozer.snoozes.values()) == [Snooze(QualifiedTargetType.DELIVERY, RecipientType.EVERYONE, "foo")]
-    assert all(s.snooze_until is not None and s.snooze_until - s.snoozed_at == 33 for s in uut.context.snoozer.snoozes.values())
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_DELIVERY_foo_33"}))
+    assert list(uut.context.snoozer.snoozes.values()) == [Snooze(
+        QualifiedTargetType.DELIVERY, RecipientType.EVERYONE, "foo")]
+    assert all(s.snooze_until is not None and s.snooze_until - s.snoozed_at == timedelta(minutes=33)
+               for s in uut.context.snoozer.snoozes.values())
 
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_NORMAL_EVERYONE_DELIVERY_foo"}))
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_NORMAL_EVERYONE_DELIVERY_foo"}))
     assert list(uut.context.snoozer.snoozes.values()) == []
 
 
 def test_snooze_everything(mock_hass: HomeAssistant) -> None:
     uut = SupernotifyAction(mock_hass)
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_EVERYTHING"}))
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_EVERYTHING"}))
     assert list(uut.context.snoozer.snoozes.values()) == [
-        Snooze(GlobalTargetType.EVERYTHING, recipient_type=RecipientType.EVERYONE)
+        Snooze(GlobalTargetType.EVERYTHING,
+               recipient_type=RecipientType.EVERYONE)
     ]
     assert all(
-        s.target is None and s.snooze_until is not None and s.snooze_until - s.snoozed_at == 3600
+        s.target is None and s.snooze_until is not None and s.snooze_until -
+        s.snoozed_at == timedelta(hours=1)
         for s in uut.context.snoozer.snoozes.values()
     )
 
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_NORMAL_EVERYONE_EVERYTHING"}))
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_NORMAL_EVERYONE_EVERYTHING"}))
     assert list(uut.context.snoozer.snoozes.values()) == []
 
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_EVERYTHING_99"}))
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_EVERYTHING_99"}))
     assert list(uut.context.snoozer.snoozes.values()) == [
-        Snooze(GlobalTargetType.EVERYTHING, recipient_type=RecipientType.EVERYONE)
+        Snooze(GlobalTargetType.EVERYTHING,
+               recipient_type=RecipientType.EVERYONE)
     ]
     assert all(
-        s.target is None and s.snooze_until is not None and s.snooze_until - s.snoozed_at == 99
+        s.target is None and s.snooze_until is not None and s.snooze_until -
+        s.snoozed_at == timedelta(minutes=99)
         for s in uut.context.snoozer.snoozes.values()
     )
 
@@ -86,15 +106,19 @@ async def test_snooze_everything_for_person(hass: HomeAssistant) -> None:
     uut = SupernotifyAction(
         hass,
         recipients=[
-            {CONF_PERSON: "person.bob_mctest", CONF_EMAIL: "bob@mctest.com", ATTR_USER_ID: "eee999111"},
-            {CONF_PERSON: "person.jane_macunit", CONF_EMAIL: "jane@macunit.org", ATTR_USER_ID: "fff444222"},
+            {CONF_PERSON: "person.bob_mctest",
+                CONF_EMAIL: "bob@mctest.com", ATTR_USER_ID: "eee999111"},
+            {CONF_PERSON: "person.jane_macunit",
+                CONF_EMAIL: "jane@macunit.org", ATTR_USER_ID: "fff444222"},
         ],
         deliveries=DELIVERY,
     )
     await uut.initialize()
-    register_mobile_app(uut.context.people_registry, person="person.bob_mctest")
+    register_mobile_app(uut.context.people_registry,
+                        person="person.bob_mctest")
     plain_notify = Notification(uut.context, "hello")
-    delivery = Delivery("email", DELIVERY["email"], uut.context.delivery_registry.transports["email"])
+    delivery = Delivery(
+        "email", DELIVERY["email"], uut.context.delivery_registry.transports["email"])
     await plain_notify.initialize()
     assert plain_notify.generate_recipients(delivery)[0].email == [
         "bob@mctest.com",
@@ -102,16 +126,20 @@ async def test_snooze_everything_for_person(hass: HomeAssistant) -> None:
     ]
 
     uut.on_mobile_action(
-        Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SNOOZE_USER_EVERYTHING"}, context=Context(user_id="eee999111"))
+        Event("mobile_action", data={
+              ATTR_ACTION: "SUPERNOTIFY_SNOOZE_USER_EVERYTHING"}, context=Context(user_id="eee999111"))
     )
     assert list(uut.context.snoozer.snoozes.values()) == [
-        Snooze(GlobalTargetType.EVERYTHING, recipient_type=RecipientType.USER, recipient="person.bob_mctest")
+        Snooze(GlobalTargetType.EVERYTHING,
+               recipient_type=RecipientType.USER, recipient="person.bob_mctest")
     ]
     await plain_notify.initialize()
-    assert plain_notify.generate_recipients(delivery)[0].email == ["jane@macunit.org"]
+    assert plain_notify.generate_recipients(delivery)[0].email == [
+        "jane@macunit.org"]
 
     uut.on_mobile_action(
-        Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_NORMAL_USER_EVERYTHING"}, context=Context(user_id="eee999111"))
+        Event("mobile_action", data={
+              ATTR_ACTION: "SUPERNOTIFY_NORMAL_USER_EVERYTHING"}, context=Context(user_id="eee999111"))
     )
     assert list(uut.context.snoozer.snoozes.values()) == []
     await plain_notify.initialize()
@@ -125,12 +153,15 @@ async def test_snooze_everything_for_person(hass: HomeAssistant) -> None:
 
 def test_clear_snoozes(mock_hass: HomeAssistant) -> None:
     uut = SupernotifyAction(mock_hass)
-    uut.on_mobile_action(Event("mobile_action", data={ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_EVERYTHING"}))
+    uut.on_mobile_action(Event("mobile_action", data={
+                         ATTR_ACTION: "SUPERNOTIFY_SNOOZE_EVERYONE_EVERYTHING"}))
     assert list(uut.context.snoozer.snoozes.values()) == [
-        Snooze(GlobalTargetType.EVERYTHING, recipient_type=RecipientType.EVERYONE)
+        Snooze(GlobalTargetType.EVERYTHING,
+               recipient_type=RecipientType.EVERYONE)
     ]
     assert all(
-        s.target is None and s.snooze_until is not None and s.snooze_until - s.snoozed_at == 3600
+        s.target is None and s.snooze_until is not None and s.snooze_until -
+        s.snoozed_at == timedelta(hours=1)
         for s in uut.context.snoozer.snoozes.values()
     )
     uut.clear_snoozes()